init experiment model
