+++ core/core/api/__init__.py
@@ -1,6 +1,14 @@
 import fastapi
 
-from . import feature_set, message_bus, project, stream_processor, target, task_runner
+from . import (
+    experiment,
+    feature_set,
+    message_bus,
+    project,
+    stream_processor,
+    target,
+    task_runner,
+)
 
 router = fastapi.APIRouter()
 router.include_router(feature_set.router, prefix="/feature-set", tags=["feature-set"])
@@ -11,3 +19,4 @@ router.include_router(
 )
 router.include_router(target.router, prefix="/target", tags=["target"])
 router.include_router(task_runner.router, prefix="/task-runner", tags=["task-runner"])
+router.include_router(experiment.router, prefix="/experiment", tags=["experiment"])
