+++ core/core/infra/__init__.py
@@ -5,3 +5,5 @@ from .message_bus import MessageBus
 from .message_bus import Message
 from .stream_processor import StreamProcessor
 from .stream_processor import SQLiteStreamProcessor
+from .task_runner import TaskRunner
+from .task_runner import SynchronousTaskRunner
